+++ neoforge/src/main/java/net/satisfy/herbalbrews/neoforge/rei/HerbalBrewsReiClientPluginNeoForge.java
@@ -1,4 +1,4 @@
-package net.satisfy.herbalbrews.forge.rei;
+package net.satisfy.herbalbrews.neoforge.rei;
 
 import me.shedaniel.rei.api.client.plugins.REIClientPlugin;
 import me.shedaniel.rei.api.client.registry.category.CategoryRegistry;
